vkconfig3: Notify no 32bit support
