+++ vkconfig_core/is_dll_32.cpp
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2020-2024 Valve Corporation
+ * Copyright (c) 2020-2024 LunarG, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Authors:
+ * - Christophe Riccio <christophe@lunarg.com>
+ */
+
+#include "is_dll_32.h"
+#include "type_platform.h"
+
+#if VKC_ENV == VKC_ENV_UNIX
+#include <unistd.h>
+
+#elif VKC_ENV == VKC_ENV_WIN32
+#include <windows.h>
+
+// From Stack Overflow.
+#define MKPTR(p1, p2) ((DWORD_PTR)(p1) + (DWORD_PTR)(p2))
+
+typedef enum _pe_architecture {
+    PE_ARCHITECTURE_UNKNOWN = 0x0000,
+    PE_ARCHITECTURE_ANYCPU = 0x0001,
+    PE_ARCHITECTURE_X86 = 0x010B,
+    PE_ARCHITECTURE_x64 = 0x020B
+} PE_ARCHITECTURE;
+
+static PE_ARCHITECTURE GetImageArchitecture(void *pImageBase) {
+    // Parse and validate the DOS header
+    IMAGE_DOS_HEADER *pDosHd = (IMAGE_DOS_HEADER *)pImageBase;
+    if (IsBadReadPtr(pDosHd, sizeof(pDosHd->e_magic)) || pDosHd->e_magic != IMAGE_DOS_SIGNATURE) return PE_ARCHITECTURE_UNKNOWN;
+
+    // Parse and validate the NT header
+    IMAGE_NT_HEADERS *pNtHd = (IMAGE_NT_HEADERS *)MKPTR(pDosHd, pDosHd->e_lfanew);
+    if (IsBadReadPtr(pNtHd, sizeof(pNtHd->Signature)) || pNtHd->Signature != IMAGE_NT_SIGNATURE) return PE_ARCHITECTURE_UNKNOWN;
+
+    // First, naive, check based on the 'Magic' number in the Optional Header.
+    PE_ARCHITECTURE architecture = (PE_ARCHITECTURE)pNtHd->OptionalHeader.Magic;
+
+    return architecture;
+}
+
+// End Stack Overflow
+#endif  // VKC_ENV
+
+/// Utility function to see if the file is 32-bit
+bool IsDLL32Bit(const Path &full_path) {
+#if VKC_ENV == VKC_ENV_WIN32
+    if (full_path.Empty()) {
+        return false;
+    }
+
+    QFile file(full_path.AbsolutePath().c_str());
+    if (!file.open(QIODevice::ReadOnly)) {
+        return false;  // punt...
+    }
+
+    // Not gonna lie, just guessed 1024 and it was enough.
+    // This is the minimum page size on any OS (I might be wrong,
+    // it could be 4096), so going lower is not really helpful.
+    // Changed to 4096 on a whim; looking for crashing bug I can't debug
+    void *header = file.map(0, 4096, QFileDevice::MapPrivateOption);
+
+    // Another punt as we may not be able to map the file
+    if (header == nullptr) {
+        return false;
+    }
+
+    PE_ARCHITECTURE arch = GetImageArchitecture(header);
+
+    file.unmap((uchar *)header);
+    file.close();
+
+    return (PE_ARCHITECTURE_X86 == arch);
+#else
+    (void)full_path;
+    return false;
+#endif  // VKC_ENV
+}
