+++ vkconfig_core/type_hide_message.h
@@ -34,9 +34,10 @@ enum HideMessageType {
     HIDE_MESSAGE_WARN_NO_LAYER_FOUND,
     HIDE_MESSAGE_WARN_MISSING_LAYERS_IGNORE,
     HIDE_MESSAGE_WARN_CORE_SHADER_IGNORE,
+    HIDE_MESSAGE_ERROR_32BIT,
 
     HIDE_MESSAGE_FIRST = HIDE_MESSAGE_NEED_APPLICATION_RESTART,
-    HIDE_MESSAGE_LAST = HIDE_MESSAGE_WARN_CORE_SHADER_IGNORE,
+    HIDE_MESSAGE_LAST = HIDE_MESSAGE_ERROR_32BIT,
 
     HIDE_MESSAGE_INVALID = ~0,
 };
