+++ vkconfig_gui/vkconfig.pro
@@ -28,6 +28,7 @@ linux: QMAKE_CXXFLAGS += -Wunused-variable
 
 SOURCES += \
     ../vkconfig_core/alert.cpp \
+    ../vkconfig_core/is_dll_32.cpp \
     ../vkconfig_core/application_singleton.cpp \
     ../vkconfig_core/bitfield.cpp \
     ../vkconfig_core/command_line.cpp \
@@ -106,6 +107,7 @@ SOURCES += \
 
 HEADERS += \
     ../vkconfig_core/alert.h \
+    ../vkconfig_core/is_dll_32.h \
     ../vkconfig_core/application_singleton.h \
     ../vkconfig_core/bitfield.h \
     ../vkconfig_core/command_line.h \
