+++ vkconfig_gui/widget_tab_configurations_layer.cpp
@@ -24,75 +24,6 @@
 
 #include "../vkconfig_core/configurator.h"
 
-#if VKC_ENV == VKC_ENV_UNIX
-#include <unistd.h>
-
-#elif VKC_ENV == VKC_ENV_WIN32
-#include <windows.h>
-
-// From Stack Overflow.
-#define MKPTR(p1, p2) ((DWORD_PTR)(p1) + (DWORD_PTR)(p2))
-
-typedef enum _pe_architecture {
-    PE_ARCHITECTURE_UNKNOWN = 0x0000,
-    PE_ARCHITECTURE_ANYCPU = 0x0001,
-    PE_ARCHITECTURE_X86 = 0x010B,
-    PE_ARCHITECTURE_x64 = 0x020B
-} PE_ARCHITECTURE;
-
-static PE_ARCHITECTURE GetImageArchitecture(void *pImageBase) {
-    // Parse and validate the DOS header
-    IMAGE_DOS_HEADER *pDosHd = (IMAGE_DOS_HEADER *)pImageBase;
-    if (IsBadReadPtr(pDosHd, sizeof(pDosHd->e_magic)) || pDosHd->e_magic != IMAGE_DOS_SIGNATURE) return PE_ARCHITECTURE_UNKNOWN;
-
-    // Parse and validate the NT header
-    IMAGE_NT_HEADERS *pNtHd = (IMAGE_NT_HEADERS *)MKPTR(pDosHd, pDosHd->e_lfanew);
-    if (IsBadReadPtr(pNtHd, sizeof(pNtHd->Signature)) || pNtHd->Signature != IMAGE_NT_SIGNATURE) return PE_ARCHITECTURE_UNKNOWN;
-
-    // First, naive, check based on the 'Magic' number in the Optional Header.
-    PE_ARCHITECTURE architecture = (PE_ARCHITECTURE)pNtHd->OptionalHeader.Magic;
-
-    return architecture;
-}
-
-// End Stack Overflow
-#endif  // VKC_ENV
-
-/// Utility function to see if the file is 32-bit
-static bool IsDLL32Bit(const std::string full_path) {
-#if VKC_ENV == VKC_ENV_WIN32
-    if (full_path.empty()) {
-        return false;
-    }
-
-    QFile file(full_path.c_str());
-    if (!file.open(QIODevice::ReadOnly)) {
-        return false;  // punt...
-    }
-
-    // Not gonna lie, just guessed 1024 and it was enough.
-    // This is the minimum page size on any OS (I might be wrong,
-    // it could be 4096), so going lower is not really helpful.
-    // Changed to 4096 on a whim; looking for crashing bug I can't debug
-    void *header = file.map(0, 4096, QFileDevice::MapPrivateOption);
-
-    // Another punt as we may not be able to map the file
-    if (header == nullptr) {
-        return false;
-    }
-
-    PE_ARCHITECTURE arch = GetImageArchitecture(header);
-
-    file.unmap((uchar *)header);
-    file.close();
-
-    return (PE_ARCHITECTURE_X86 == arch);
-#else
-    (void)full_path;
-    return false;
-#endif  // VKC_ENV
-}
-
 ConfigurationLayerWidget::ConfigurationLayerWidget(TabConfigurations *tab, const Parameter &parameter,
                                                    const std::vector<Version> &layer_versions, bool advanced_view)
     : tab(tab), layer_name(parameter.key) {
@@ -134,19 +65,19 @@ ConfigurationLayerWidget::ConfigurationLayerWidget(TabConfigurations *tab, const
             this->setToolTip(layer->description.c_str());
         }
 
-        // this->layer_version->setEnabled(layers.size() > 1);
         this->connect(this->layer_version, SIGNAL(currentIndexChanged(int)), this, SLOT(on_layer_version_currentIndexChanged(int)));
-        // this->layer_version->installEventFilter(this);
 
         this->layer_state = new QComboBox(this);
         if (layer != nullptr) {
             if (layer->type == LAYER_TYPE_EXPLICIT) {
                 for (int i = LAYER_CONTROL_EXPLICIT_FIRST; i <= LAYER_CONTROL_EXPLICIT_LAST; ++i) {
                     this->layer_state->addItem(GetToken(static_cast<LayerControl>(i)));
+                    this->layer_state->setItemData(i, GetDescription(static_cast<LayerControl>(i)), Qt::ToolTipRole);
                 }
             } else {
                 for (int i = LAYER_CONTROL_IMPLICIT_FIRST; i <= LAYER_CONTROL_IMPLICIT_LAST; ++i) {
                     this->layer_state->addItem(GetToken(static_cast<LayerControl>(i)));
+                    this->layer_state->setItemData(i, GetDescription(static_cast<LayerControl>(i)), Qt::ToolTipRole);
                 }
             }
         }
@@ -155,7 +86,6 @@ ConfigurationLayerWidget::ConfigurationLayerWidget(TabConfigurations *tab, const
         this->layer_state->setToolTip(GetDescription(parameter.control));
         this->layer_state->setEnabled(layer != nullptr);
         this->connect(this->layer_state, SIGNAL(currentIndexChanged(int)), this, SLOT(on_layer_state_currentIndexChanged(int)));
-        // this->layer_state->installEventFilter(this);
     }
 
     std::string decorated_name = parameter.key;
@@ -180,10 +110,6 @@ ConfigurationLayerWidget::ConfigurationLayerWidget(TabConfigurations *tab, const
         if (layer->status != STATUS_STABLE) {
             decorated_name += format(" (%s)", GetToken(layer->status));
         }
-
-        // if (IsDLL32Bit(layer->manifest_path)) {
-        //    decorated_name += " (32-bit)";
-        //}
     }
 
     this->setText((" =  " + decorated_name).c_str());
